Module type Append_only.Maker

Parameters

module K : Type.S

module V : Type.S

Signature

Append-only stores

Append-only stores are store where it is possible to read and add new values.

Read-only stores

Read-only stores are store where it is only possible to read existing values.

type -'a t

The type for stores. The 'a phantom type carries information about the store mutability.

type key = K.t

The type for keys.

type value = V.t

The type for raw values.

val mem : [> Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Perms.write ] t -> key -> value -> unit Lwt.t

Write the contents of a value to the store.

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

val batch : Perms.read t -> ( [ Perms.read | Perms.write ] t -> 'a Lwt.t ) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depend on the implementation.

val v : Irmin__Conf.t -> Perms.read t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.