Module Irmin.Backend

Backend defines functions only useful for creating new backends. If you are just using the library (and not developing a new backend), you should not use this module.

module Conf : sig ... end

Backend configuration.

module Watch : sig ... end

Watch provides helpers to register event notifications on read-write stores.

module Lock : sig ... end

module Lru : sig ... end

module Slice : sig ... end

module Remote : sig ... end

Remote stores.

module type S = Backend.S

The modules that define a complete Irmin backend. Apply an implementation to Of_backend to create an Irmin store.