Parameter Store.3-C

Signature for store contents

type t = S.value

The type for user-defined contents.

val t : t Type.t

val merge : t option Merge.t

Merge function. Evaluates to `Conflict msg if the values cannot be merged properly. The arguments of the merge function can take None to mean that the key does not exists for either the least-common ancestor or one of the two merging points. The merge function returns None when the key's value should be deleted.