Module Irmin.Export_for_backends

Helper module containing useful top-level types for defining Irmin backends. This module is relatively unstable.

module Store_properties : sig ... end

module Logging : sig ... end

Logs tags attached to the log entries emitted by Irmin:

module Reversed_list : sig ... end

Reversed_list.t is constructed the same way as List.t, but needs to be reversed before it can be used as a regular list.

type read = Perms.read

type write = Perms.write

type read_write = Perms.read_write

Lwt syntax

include module type of struct include Lwt.Syntax end

val let* : 'a Lwt.t -> ('a -> 'b Lwt.t) -> 'b Lwt.t

val and* : 'a Lwt.t -> 'b Lwt.t -> ('a * 'b) Lwt.t

val let+ : 'a Lwt.t -> ('a -> 'b) -> 'b Lwt.t

val and+ : 'a Lwt.t -> 'b Lwt.t -> ('a * 'b) Lwt.t

val (>>=) : 'a Lwt.t -> ('a -> 'b Lwt.t) -> 'b Lwt.t

val (>|=) : 'a Lwt.t -> ('a -> 'b) -> 'b Lwt.t

Dependency extensions

module Option : sig ... end

module List : sig ... end

module Seq : sig ... end

val shuffle : Stdlib.Random.State.t -> 'a array -> unit