Module type Store_properties.Of_config

type 'a t

val v : Irmin__.Conf.t -> Perms.read t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.