Module Irmin.Generic_key

"Generic key" stores are Irmin stores in which the backend may not be keyed directly by the hashes of stored values. See Key for more details.

module type S = Irmin__.Store_intf.S_generic_key

module type KV = Irmin__.Store_intf.KV_generic_key

module type Maker = Irmin__.Store_intf.Maker_generic_key

module type KV_maker = Irmin__.Store_intf.KV_maker_generic_key

module type Maker_args = sig ... end

module Maker : functor (X : Maker_args) -> Maker with type ('h, 'v) contents_key = ('h, 'v) X.Contents_store.key and type 'h node_key = 'h X.Node_store.key and type 'h commit_key = 'h X.Commit_store.key