Module Hash.Typed

Typed hashes.

Parameters

module K : S

module V : Type.S

Signature

type t = K.t

type value = V.t

val hash : value -> t

Compute a deterministic store key from a string.

val short_hash : t -> int

short_hash h is a small hash of h, to be used for instance as the `hash` function of an OCaml Hashtbl.

val hash_size : int

hash_size is the size of hash results, in bytes.

Value Types

val t : t Type.t

t is the value type for t.