Module Hash.V1

v1 serialisation

Parameters

module H : S

Signature

type t = H.t

The type for digest hashes.

val hash : ((string -> unit) -> unit) -> t

Compute a deterministic store key from a sequence of strings.

val short_hash : t -> int

short_hash h is a small hash of h, to be used for instance as the `hash` function of an OCaml Hashtbl.

val hash_size : int

hash_size is the size of hash results, in bytes.

Value Types

val t : t Type.t

t is the value type for t.