Module type Hash.TYPED

Signature for typed hashes, where hash directly takes a value as argument and incremental hashing is not possible.

type t

type value

val hash : value -> t

Compute a deterministic store key from a string.

val short_hash : t -> int

short_hash h is a small hash of h, to be used for instance as the `hash` function of an OCaml Hashtbl.

val hash_size : int

hash_size is the size of hash results, in bytes.

Value Types

val t : t Type.t

t is the value type for t.