Module Irmin.Indexable

Indexable backend backends.

module type S = sig ... end

An indexable store is a read-write store in which values can be added and later found via their keys.

module type S_without_key_impl = sig ... end

module type Maker = sig ... end

module type Maker_concrete_key1 = sig ... end

module type Maker_concrete_key2 = sig ... end

module Maker_concrete_key2_of_1 (X : Maker_concrete_key1) : Maker_concrete_key2 with type ('h, _) key = 'h X.key

module Of_content_addressable (Key : Type.S) (S : Content_addressable.S with type key = Key.t) : S with type 'a t = 'a S.t and type key = Key.t and type hash = Key.t and type value = S.value

An indexable store is a read-write store in which values can be added and later found via their keys.

module Check_closed_store (CA : S) : sig ... end

module Check_closed (M : Maker) : Maker