Module Irmin.Info

Commit info are used to keep track of the origin of write operations in the stores. Info models the metadata associated with commit objects in Git.

Commit Info

type t

The type for commit info.

val v : date:int64 -> author:string -> string -> t

Create a new commit info.

val date : t -> int64

date t is t's commit date.

The date provided by the user when calling the create function. Rounding Unix.gettimeofday () (when available) is a good value for such date. On more esoteric platforms, any monotonic counter is a fine value as well. On the Git backend, the date is translated into the commit Date field and is expected to be the number of POSIX seconds (thus not counting leap seconds) since the Epoch.

val author : t -> string

author t is t's commit author.

The author identifies the entity (human, unikernel, process, thread, etc) performing an operation. For the Git backend, this will be directly translated into the Author field.

val message : t -> string

message t is t's commit message.

val empty : t

The empty commit info.

Info Functions

type f = unit -> t

Alias for functions which can build commit info.

val none : f

The empty info function. none () is empty

Value Types

val t : t Type.t

t is the value type for t.