Module Node.Generic_key

Generic_key generalises the concept of "node" to one that supports object keys that are not strictly equal to hashes.

module type S = sig ... end

module type Maker = sig ... end

module type Core = sig ... end

module Make : Maker

module Make_v2 : Maker

Make_v2 provides a similar implementation as Make but the hash computation is compatible with versions older than irmin.3.0

module Store (C : Contents.Store) (S : Indexable.S) (H : Hash.S with type t = S.hash) (V : S with type t = S.value and type hash = H.t and type contents_key = C.key and type node_key = S.key) (M : Metadata.S with type t = V.metadata) (P : Path.S with type step = V.step) : sig ... end