Module Node.Val

Val provides base functions for node values.

Node values

type t = value

The type for node values.

type metadata = Metadata.t

The type for node metadata.

type hash = key

The type for keys.

type step = Path.step

The type for steps between nodes.

type value = [

| `Node of hash

| `Contents of hash * metadata

]

The type for either (node) keys or (contents) keys combined with their metadata.

val v : (step * value) list -> t

create l is a new node.

val list : ?⁠offset:int -> ?⁠length:int -> t -> (step * value) list

list t is the contents of t. offset and length are used to paginate results.

val empty : t

empty is the empty node.

val is_empty : t -> bool

is_empty t is true iff t is empty.

val find : t -> step -> value option

find t s is the value associated with s in t.

A node can point to user-defined contents. The edge between the node and the contents is labeled by a step.

val add : t -> step -> value -> t

add t s v is the node where find t v is Some s but is similar to t otherwise.

val remove : t -> step -> t

remove t s is the node where find t s is None but is similar to t otherwise.

Value types

val t : t Type.t

t is the value type for t.

val default : metadata

default is the default metadata value.

val metadata_t : metadata Type.t

metadata_t is the value type for metadata.

val hash_t : hash Type.t

hash_t is the value type for hash.

val step_t : step Type.t

step_t is the value type for step.

val value_t : value Type.t

value_t is the value type for value.