Module Irmin.Private

Private defines functions only useful for creating new backends. If you are just using the library (and not developing a new backend), you should not use this module.

module Conf : sig ... end

Backend configuration.

module Watch : sig ... end

Watch provides helpers to register event notifications on read-write stores.

module Lock : sig ... end

module Lru : sig ... end

module Node : sig ... end

Node provides functions to describe the graph-like structured values.

module Commit : sig ... end

Commit values represent the store history.

module Slice : sig ... end

module Sync = Sync

module Sigs : sig ... end

Irmin signatures

module type S = Private.S

The complete collection of private implementations.