Module Node.Store

Store creates node stores.

Parameters

module C : Contents.STORE

module P : Path.S

module M : sig ... end

module N : sig ... end

Signature

type 'a t = 'a C.t * 'a N.t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key = N.key

The type for keys.

type value = N.value

The type for raw values.

val mem : [> Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Perms.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Perms.write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

module Path = P

Path provides base functions on node paths.

val merge : [> Perms.read_write ] t -> key option Merge.t

merge is the 3-way merge function for nodes keys.

module Key : Hash.TYPED with type t = key and type value = value with type t = N.key

Key provides base functions for node keys.

module Metadata = M

Metadata provides base functions for node metadata.

module Val = N.Val

Val provides base functions for node values.

module Contents : Contents.STORE with type key = Val.hash

Contents is the underlying contents store.