Module 4-N.Val

Node values

type t = value

The type for node values.

val t : t Type.t

type metadata = M.t

The type for node metadata.

val metadata_t : metadata Type.t

type hash = key

The type for keys.

val hash_t : hash Type.t

type step = P.step

The type for steps between nodes.

val step_t : step Type.t

type value = [

| `Node of hash

| `Contents of hash * metadata

]

The type for either (node) keys or (contents) keys combined with their metadata.

val value_t : value Type.t

val of_list : (step * value) list -> t

of_list l is the node n such that list n = l.

val list : ?offset:int -> ?length:int -> ?cache:bool -> t -> (step * value) list

list t is the contents of t. offset and length are used to paginate results.

caching

cache regulates the caching behaviour regarding the node's internal data which may be lazily loaded from the backend, depending on the node implementation.

cache defaults to true which may greatly reduce the IOs and the runtime but may also increase the memory consumption.

cache = false doesn't replace a call to clear, it only prevents the storing of new data, it doesn't discard the existing one.

val of_seq : ( unit -> (step * value) Stdlib__seq.node ) -> t

of_seq s is the node n such that seq n = s.

val seq : ?offset:int -> ?length:int -> ?cache:bool -> t -> unit -> (step * value) Stdlib__seq.node

seq t is the contents of t. offset and length are used to paginate results.

See caching for an explanation of the cache parameter

val empty : t

empty is the empty node.

val is_empty : t -> bool

is_empty t is true iff t is empty.

val length : t -> int

length t is the number of entries in t.

val clear : t -> unit

Cleanup internal caches.

val find : ?cache:bool -> t -> step -> value option

find t s is the value associated with s in t.

A node can point to user-defined contents. The edge between the node and the contents is labeled by a step.

See caching for an explanation of the cache parameter

val add : t -> step -> value -> t

add t s v is the node where find t v is Some s but is similar to t otherwise.

val remove : t -> step -> t

remove t s is the node where find t s is None but is similar to t otherwise.

val default : metadata

default is the default metadata value.

Proofs

type nonrec proof

The type for proof trees.

val proof_t : proof Type.t

val to_proof : t -> proof

val of_proof : proof -> t