Module type Sigs.APPEND_ONLY_STORE_MAKER

Parameters

module K : Type.S

module V : Type.S

Signature

include APPEND_ONLY_STORE with type key = K.t and type value = V.t

type -'a t

The type for append-only backend stores. The 'a phantom type carries information about the store mutability.

type key = K.t

The type for keys.

type value = V.t

The type for raw values.

val mem : [> Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Perms.write ] t -> key -> value -> unit Lwt.t

Write the contents of a value to the store.

include Store_properties.CLEARABLE with type 'a t := 'a t

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

include Store_properties.BATCH with type 'a t := 'a t

val batch : Perms.read t -> ([ Perms.read | Perms.write ] t -> 'a Lwt.t) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depend on the implementation.

include Store_properties.OF_CONFIG with type 'a t := 'a t

val v : config -> Perms.read t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.

include Store_properties.CLOSEABLE with type 'a t := 'a t

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.