Module type Irmin.APPEND_ONLY_STORE

Append-only backend store.

type -'a t

The type for append-only backend stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Perms.write ] t -> key -> value -> unit Lwt.t

Write the contents of a value to the store.

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.