Module type Irmin.Maker

Maker is the signature exposed by any backend providing S implementations. M is the implementation of user-defined metadata, C is the one for user-defined contents, B is the implementation for branches and H is the implementation for object (blobs, trees, commits) hashes. It does not use any native synchronization primitives.

type endpoint

include Key.Store_spec.S with type ('h, _) contents_key = 'h with type 'h node_key = 'h with type 'h commit_key = 'h

type ('h, _) contents_key = 'h

type 'h node_key = 'h

type 'h commit_key = 'h

module Make (Schema : Schema.S) : sig ... end