Module type Append_only_intf.Maker

Parameters

K : Irmin__.Type.S

V : Irmin__.Type.S

Signature

include S with type key = K.t and type value = V.t

type -'a t

The type for stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin__.Import.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin__.Import.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

type 'a t

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

val add : [> Irmin__.Import.write ] t -> key -> value -> unit Lwt.t

Write the contents of a value to the store.

type 'a t

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

type 'a t

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

type 'a t

val batch : Irmin__.Import.read t -> ([ Irmin__.Import.read | Irmin__.Import.write ] t -> 'a Lwt.t) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depend on the implementation.

type 'a t

val v : Irmin__.Conf.t -> Irmin__.Import.read t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.