Module type Commit_intf.Sigs

module type S = S

module type Maker = Maker

module Maker : functor (I : Irmin.Info.S) -> Maker with module Info = I

Maker provides a simple implementation of commit values, parameterized by commit info.

module Generic_key : sig ... end

Generic_key generalises the concept of "commit" to one that supports object keys that are not strictly equal to hashes.

module V1 : sig ... end

V1 serialisation.

module Portable : sig ... end

module type Store = Store

Store specifies the signature for commit stores.

module Store : functor (I : Irmin.Info.S) -> functor (N : Irmin.Node.Store) -> functor (S : Irmin.Content_addressable.S with type key = N.key) -> functor (H : Irmin.Hash.S with type t = S.key) -> functor (V : S with type hash = S.key and type t = S.value and module Info := I) -> Store with type 'a t = 'a N.t * 'a S.t and type key = S.key and type hash = S.key and type value = S.value and module Info = I and module Val = V

Store creates a new commit store.

module type History = History

History specifies the signature for commit history. The history is represented as a partial-order of commits and basic functions to search through that history are provided.

module History : functor (C : Store) -> History with type 'a t = 'a C.t and type v = C.Val.t and type node_key = C.Node.key and type commit_key = C.key and type info = C.Info.t

Build a commit history.

include Maker with module Maker.Info = Irmin.Info.Default

module Info : sig ... end

module Make : functor (H : Irmin__.Type.S) -> sig ... end