Module type Hash_intf.Sigs

module type S = S

Signature for hash values.

module type Typed = Typed

Signature for typed hashes, where hash directly takes a value as argument and incremental hashing is not possible.

module Make : functor (H : Digestif.S) -> S with type t = H.t

Digestif hashes.

module Make_BLAKE2B : functor (D : sig ... end) -> S

module Make_BLAKE2S : functor (D : sig ... end) -> S

module SHA1 : S

module RMD160 : S

module SHA224 : S

module SHA256 : S

module SHA384 : S

module SHA512 : S

module BLAKE2B : S

module BLAKE2S : S

module V1 : functor (H : S) -> S with type t = H.t

v1 serialisation

module Typed : functor (K : S) -> functor (V : Irmin__.Type.S) -> Typed with type t = K.t and type value = V.t

Typed hashes.