Module Irmin__.Indexable_intf

module type S_without_key_impl = sig ... end

module type S = sig ... end

module type Maker = functor (Hash : Irmin.Hash.S) -> functor (Value : Irmin__.Type.S) -> sig ... end

module type Maker_concrete_key1 = sig ... end

A Maker_concrete_key is an indexable store in which the key type is uniquely determined by the hash type and is stated up-front.

module type Maker_concrete_key2 = sig ... end

Like Maker_concrete_key1, but the key type may also depend on type of the value that it references.

module type Sigs = sig ... end