Module type Key_intf.Sigs

module type S = S

module type Hash_like = Hash_like

module Of_hash : functor (H : Irmin__.Type.S) -> Hash_like with type t = H.t and type hash = H.t

The simplest possible Key implementation is just a hash of the corresponding value, attaching no additional metadata about the value.

module Store_spec : sig ... end