Module Sigs.Of_hash

The simplest possible Key implementation is just a hash of the corresponding value, attaching no additional metadata about the value.

Parameters

H : Irmin__.Type.S

Signature

include S

type t

The type for keys.

val t : t Irmin__.Type.t

type hash

val to_hash : t -> hash

val of_hash : hash -> t