Module Node.Generic_key

Generic_key generalises the concept of "node" to one that supports object keys that are not strictly equal to hashes.

module type S = Irmin__.Node_intf.S_generic_key

module type Maker = Irmin__.Node_intf.Maker_generic_key

module type Core = Irmin__.Node_intf.Core

module Make : Maker

module Make_v2 : Maker

Make_v2 provides a similar implementation as Make but the hash computation is compatible with versions older than irmin.3.0

module Store : functor (C : Irmin.Contents.Store) -> functor (S : Irmin.Indexable.S) -> functor (H : Irmin.Hash.S with type t = S.hash) -> functor (V : S with type t = S.value and type hash = H.t and type contents_key = C.key and type node_key = S.key) -> functor (M : Irmin.Metadata.S with type Store.t = V.metadata) -> functor (P : Irmin.Path.S with type step = V.step) -> Irmin__.Node_intf.Store with type 'a t = 'a C.t * 'a S.t and type key = S.key and type hash = S.hash and type value = S.value and module Path = P and module Metadata = M and module Val = V