Module type Node_intf.Node

module type S = S

module type Maker = Maker

module Make : Maker

Make provides a simple node implementation, parameterized by the contents and notes keys K, paths P and metadata M.

module V1 : functor (N : S with type step = string) -> sig ... end

v1 serialisation

module type STORE = STORE

STORE specifies the signature for node stores.

module Store : functor (C : Irmin.Contents.STORE) -> functor (P : Irmin.Path.S) -> functor (M : Irmin__.S.METADATA) -> functor (N : sig ... end) -> STORE with type 'a t = 'a C.t * 'a N.t and type key = N.key and type value = N.value and module Path = P and module Metadata = M and type Key.t = N.key and module Val = N.Val

Store creates node stores.

module type GRAPH = GRAPH

Graph specifies the signature for node graphs. A node graph is a deterministic DAG, labeled by steps.

module Graph : functor (N : STORE) -> GRAPH with type 'a t = 'a N.t and type contents = N.Contents.key and type metadata = N.Metadata.t and type node = N.key and type step = N.Path.step and type path = N.Path.t

module No_metadata : Irmin__.S.METADATA with type t = unit