Module Irmin__.Perms

module Read : sig ... end

module Write : sig ... end

module Read_write : sig ... end

type read = Read.t

The type parameter of a handle with read permissions.

type write = Write.t

The type parameter of a handle with write permissions.

type read_write = Read_write.t

The type parameter of a handle with both read and write permissions.