Module type Read_only_intf.Maker

Parameters

Key : Irmin__.Type.S

Value : Irmin__.Type.S

Signature

include S with type key = Key.t and type value = Value.t

type -'a t

The type for stores. The 'a phantom type carries information about the store mutability.

type key = Key.t

The type for keys.

type value = Value.t

The type for raw values.

val mem : [> Irmin__.Import.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin__.Import.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

type 'a t

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

type 'a t

val v : Irmin__.Conf.t -> Irmin__.Import.read t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.