Module type Store_properties.OF_CONFIG

type 'a t

val v : config -> Irmin__.Import.read t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.