Module type Store_intf.Store

module type S = S

module type MAKER = MAKER

module type JSON_TREE = JSON_TREE

type Sigs.remote += Store : (module S with type t = 'a) * 'a -> Sigs.remote

module Make : functor (P : Irmin.Private.S) -> S with type key = P.Node.Path.t and type contents = P.Contents.value and type branch = P.Branch.key and type hash = P.Hash.t and type slice = P.Slice.t and type step = P.Node.Path.step and type metadata = P.Node.Metadata.t and module Key = P.Node.Path and type repo = P.Repo.t and module Private = P

module Json_tree : JSON_TREE

Json_tree is used to project JSON values onto trees. Instead of the entire object being stored under one key, it is split across several keys starting at the specified root key.

module Content_addressable : functor (X : Sigs.APPEND_ONLY_STORE_MAKER) -> functor (K : Irmin.Hash.S) -> functor (V : Irmin__.Type.S) -> sig ... end

module type Maker_future = Maker_future

Exported for compatibility with a future version of Irmin.