Module Make.Contents

Operations on lazy tree contents.

type t

The type of lazy tree contents.

val hash : ?⁠cache:bool -> t -> hash

hash t is the hash of the contents value returned when t is forced successfully. See caching for an explanation of the cache parameter.

val key : t -> contents_key option

key t is the key of the contents value returned when t is forced successfully.

val force : t -> contents or_error Lwt.t

force t forces evaluation of the lazy content value t, or returns an error if no such value exists in the underlying repository.

val force_exn : t -> contents Lwt.t

Equivalent to force, but raises an exception if the lazy content value is not present in the underlying repository.

val clear : t -> unit

clear t clears t's cache.

caching

cache regulates the caching behaviour regarding the node's internal data which are be lazily loaded from the backend.

cache defaults to true which may greatly reduce the IOs and the runtime but may also grealy increase the memory consumption.

cache = false doesn't replace a call to clear, it only prevents the storing of new data, it doesn't discard the existing one.