Module Irmin__Commit

Commit values represent the store history.

Every commit contains a list of predecessor commits, and the collection of commits form an acyclic directed graph.

Every commit also can contain an optional key, pointing to a node value. See the Node signature for more details on node values.