Module type Irmin__Commit_intf.Commit

module type S = S

module type Maker = Maker

module Make : Maker

Make provides a simple implementation of commit values, parameterized by the commit and node keys K.

module V1 : functor (C : S) -> sig ... end

V1 serialisation.

module type STORE = STORE

STORE specifies the signature for commit stores.

module Store : functor (N : Irmin__.Node.STORE) -> functor (C : sig ... end) -> STORE with type 'a t = 'a N.t * 'a C.t and type key = C.key and type value = C.value and type Key.t = C.Key.t and module Val = C.Val

Store creates a new commit store.

module type HISTORY = HISTORY

History specifies the signature for commit history. The history is represented as a partial-order of commits and basic functions to search through that history are provided.

module History : functor (C : STORE) -> HISTORY with type 'a t = 'a C.t and type v = C.Val.t and type node = C.Node.key and type commit = C.key

Build a commit history.