Module Irmin__Contents.Json_tree

Parameters

Store : Irmin__.S.STORE with type contents = json

Signature

include Irmin__.S.CONTENTS with type t = json

type t = json

The type for user-defined contents.

val t : t Irmin.Type.t

t is the value type for t.

val merge : t option Irmin.Merge.t

Merge function. Evaluates to `Conflict msg if the values cannot be merged properly. The arguments of the merge function can take None to mean that the key does not exists for either the least-common ancestor or one of the two merging points. The merge function returns None when the key's value should be deleted.

val to_concrete_tree : t -> Store.Tree.concrete

val of_concrete_tree : Store.Tree.concrete -> t

val get_tree : Store.tree -> Store.key -> json Lwt.t

val set_tree : Store.tree -> Store.key -> json -> Store.tree Lwt.t

val get : Store.t -> Store.key -> json Lwt.t

val set : Store.t -> Store.key -> json -> info:Irmin.Info.f -> unit Lwt.t