Module type Irmin__Contents_intf.Sigs

module type S = S

module String : S with type t = string

Contents of type string, with the default 3-way merge strategy: assume that update operations are idempotent and conflict iff values are modified concurrently.

module String_v2 : S with type t = string

Similar to String above, but the hash computation is compatible with versions older than irmin.3.0

type json = [

| `Null

| `Bool of bool

| `String of string

| `Float of float

| `O of (string * json) list

| `A of json list

]

module Json : S with type t = (string * json) list

Json contents are associations from strings to json values stored as JSON encoded strings. If the same JSON key has been modified concurrently with different values then the merge function conflicts.

module Json_value : S with type t = json

Json_value allows any kind of json value to be stored, not only objects.

module V1 : sig ... end

module type Store = Store

Contents store.

module Store : functor (S : Irmin.Content_addressable.S) -> functor (H : Irmin.Hash.S with type t = S.key) -> functor (C : S with type t = S.value) -> Store with type 'a t = 'a S.t and type key = H.t and type hash = H.t and type value = C.t

Store creates a contents store.

module Store_indexable : functor (S : Irmin.Indexable.S) -> functor (H : Irmin.Hash.S with type t = S.hash) -> functor (C : S with type t = S.value) -> Store with type 'a t = 'a S.t and type key = S.key and type value = S.value and type hash = S.hash

Store_indexable is like Store but uses an indexable store as a backend (rather than a content-addressable one).