Module type Irmin__Node_intf.Sigs

module type S = S

module Make : functor (Hash : Irmin.Hash.S) -> functor (Path : sig ... end) -> functor (Metadata : Irmin.Metadata.S) -> S with type hash = Hash.t and type metadata = Metadata.t and type step = Path.step

Make provides a simple node implementation, parameterized by hash, path and metadata implementations. The contents and node values are addressed directly by their hash.

module Generic_key : sig ... end

Generic_key generalises the concept of "node" to one that supports object keys that are not strictly equal to hashes.

module V1 : functor (N : Generic_key.S with type step = string) -> sig ... end

v1 serialisation

module Portable : sig ... end

module type Store = Store

Store specifies the signature for node stores.

module Store : functor (C : Irmin.Contents.Store) -> functor (S : Irmin.Content_addressable.S with type key = C.key) -> functor (H : Irmin.Hash.S with type t = S.key) -> functor (V : S with type t = S.value and type hash = S.key) -> functor (M : Irmin.Metadata.S with type Store.t = V.metadata) -> functor (P : Irmin.Path.S with type step = V.step) -> Store with type 'a t = 'a C.t * 'a S.t and type key = S.key and type value = S.value and type hash = H.t and module Path = P and module Metadata = M and module Val = V

Store creates node stores.

module type Graph = Graph

Graph specifies the signature for node graphs. A node graph is a deterministic DAG, labeled by steps.

module Graph : functor (N : Store) -> Graph with type 'a t = 'a N.t and type contents_key = N.Contents.key and type node_key = N.key and type metadata = N.Metadata.t and type step = N.Path.step and type path = N.Path.t