Module type Irmin__Proof_intf.Env

Environment that tracks side effects during the production/consumption of proofs.

The merkle proof construction algorithm

This description assumes that the large nodes are represented by the backend as a tree structure (i.e. inodes). There are 4 distinct phases when working with Irmin's merkle proofs: Produce | Serialise | Deserialise | Consume.

Produce

In this phase the Irmin user builds an after tree from a before tree that has been setup with an Env that records every backend reads into two hash tables.

During the next phase (i.e. Serialise) the cleared before tree will be traversed from root to stems only following the paths that are referenced in Env.

In practice Env doesn't exactly records the reads, it keeps track of all the hash -> backend node and hash -> backend contents mappings that are directly output of the backend stores through P.Node.find and P.Contents.find. This is obviously enough to remember the contents, the nodes and the inodes tips, but the inner inodes are not directly referenced in the hash tables.

The inner inodes are in fact referenced in their inode tip which is itself referenced in Env's hash tables. Since an inode shares its lazy pointers with the inodes derived from it, even the inner inodes that are loaded from the derived tips will be available from the original inode tip.

Serialise

In this phase the Env contains everything necessary for the computation of a merkle proof from a cleared before. The Env now affects Node.cached_value and Contents.cached_value allowing for the discovery of the cached closure.

Deserialise

In this phase the Env is filled by recursively destructing the proof and filling it before the Consume phase.

Consume

In this last phase the Env is again made accessible through Node.cached_value and Contents.cached_value, making it possible for the user to reference by hash everything that was contained in the proof.

type kind =

| Set

| Stream

type mode =

| Produce

| Serialise

| Deserialise

| Consume

type v

type t

type hash

type node

type contents

type stream

val t : t Irmin__.Type.ty

val is_empty : t -> bool

Construction of envs

val empty : unit -> t

val copy : into:t -> t -> unit

Modes

val mode : t -> mode option

val set_mode : t -> kind -> mode -> unit

val with_set_produce : (t -> start_serialise:(unit -> unit) -> 'a Lwt.t) -> 'a Lwt.t

val with_set_consume : (t -> stop_deserialise:(unit -> unit) -> 'a Lwt.t) -> 'a Lwt.t

val with_stream_produce : (t -> to_stream:(unit -> stream) -> 'a Lwt.t) -> 'a Lwt.t

val with_stream_consume : stream -> (t -> is_empty:(unit -> bool) -> 'a Lwt.t) -> 'a Lwt.t

In/out backend objects with Tree

val add_contents_from_store : t -> hash -> contents -> unit

val add_node_from_store : t -> hash -> node -> node

add_node_from_store returns a node and not unit because Env may take the opportunity to wrap the input node in Node.Val.with_handler.

val add_contents_from_proof : t -> hash -> contents -> unit

val add_node_from_proof : t -> hash -> node -> unit

val find_contents : t -> hash -> contents option

val find_node : t -> hash -> node option