Module type Proof.Env

type kind =

| Set

| Stream

type mode =

| Produce

| Serialise

| Deserialise

| Consume

type v

type t

type hash

type node

type contents

type stream

val t : t Irmin__.Type.ty

val is_empty : t -> bool

Construction of envs

val empty : unit -> t

val copy : into:t -> t -> unit

Modes

val mode : t -> mode option

val set_mode : t -> kind -> mode -> unit

val with_set_produce : (t -> start_serialise:(unit -> unit) -> 'a Lwt.t) -> 'a Lwt.t

val with_set_consume : (t -> stop_deserialise:(unit -> unit) -> 'a Lwt.t) -> 'a Lwt.t

val with_stream_produce : (t -> to_stream:(unit -> stream) -> 'a Lwt.t) -> 'a Lwt.t

val with_stream_consume : stream -> (t -> is_empty:(unit -> bool) -> 'a Lwt.t) -> 'a Lwt.t

In/out backend objects with Tree

val add_contents_from_store : t -> hash -> contents -> unit

val add_node_from_store : t -> hash -> node -> node

add_node_from_store returns a node and not unit because Env may take the opportunity to wrap the input node in Node.Val.with_handler.

val add_contents_from_proof : t -> hash -> contents -> unit

val add_node_from_proof : t -> hash -> node -> unit

val find_contents : t -> hash -> contents option

val find_node : t -> hash -> node option