Module Irmin__Store

Branch-consistent stores: read-write store with support fork/merge operations.

module type S = Irmin__.Store_intf.S

module type Maker = Irmin__.Store_intf.Maker

module type Json_tree = Irmin__.Store_intf.Json_tree

module type KV = Irmin__.Store_intf.KV

module type KV_maker = Irmin__.Store_intf.KV_maker

module Generic_key : sig ... end

type Irmin__.Remote.t += Store : (module Generic_key.S with type t = 'a) * 'a -> Irmin__.Remote.t

module Make : functor (B : Irmin.Backend.S) -> Generic_key.S with module Schema = B.Schema and type slice = B.Slice.t and type repo = B.Repo.t and type contents_key = B.Contents.key and type node_key = B.Node.key and type commit_key = B.Commit.key and module Backend = B

module Json_tree : Json_tree

Json_tree is used to project JSON values onto trees. Instead of the entire object being stored under one key, it is split across several keys starting at the specified root key.