Module type Sigs.Of_config

type 'a t

val v : Irmin__.Conf.t -> Irmin__.Import.read t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.