Irmin

Irmin is a distributed database built on the same principles as Git. It can be used as an OCaml library that provides persistent stores with built-in snapshot, branching and reverting mechanisms.

Release 3.5.2 - %%HOMEPAGE%%

Irmin API

The irmin package defines the public API of Irmin. The entry point of the library is the module Irmin.

Backends

Irmin is designed to use a large variety of backends. Various backends are provided:

Irmin_mem: An in-memory store provided in the irmin package (this package).

Irmin_git: A backend that exposes a bi-directional bridge to a Git repository proivded by the irmin-git package.

Irmin_pack: A space-optimiezed, on-disk store inspired by Git Packfiles is provided in the irmin-pack package.

Irmin_mirage: A MirageOS compatible store is provided in the irmin-mirage package.

Helpers and Internal Utilities

irmin.unix

The irmin.unix package provides Irmin related utilities for Unix applications.

The entry point of this library is the module Irmin_unix.

irmin.data

This package contains data structure implementations used in the implementation of Irmin. It is exposed only for internal use, and does not provide a stable API.

The entry point of this library is the module Irmin_data.