Module Conf.Spec

type t

A configuration spec is used to group keys by backend

val v : string -> t

v name is a new configuration specification named name

val name : t -> string

name spec is the name associated with a config spec

val list : unit -> t Stdlib.Seq.t

list () is a sequence containing all available config specs

val find : string -> t option

find name is the config spec associated with name if available

val find_key : t -> string -> k option

find_key spec k is the key associated with the name k in spec

val keys : t -> k Stdlib.Seq.t

keys spec is a sequence of keys available in spec

val join : t -> t list -> t

join a b is a new Spec.t combining a and all specs present in b

The name of the resulting spec will be the name of a and the names of the specs in b joined by hyphens.