Module Irmin.Dot

Parameters

module S : S

Signature

Dot Export

type db = S.t

val output_buffer : db -> ?⁠html:bool -> ?⁠depth:int -> ?⁠full:bool -> date:(int64 -> string) -> Stdlib.Buffer.t -> unit Lwt.t

output_buffer t ?html ?depth ?full buf outputs the Graphviz representation of t in the buffer buf.

html (default is false) enables HTML labels.

depth is used to limit the depth of the commit history. None here means no limitation.

If full is set (default is not) the full graph, including the commits, nodes and contents, is exported, otherwise it is the commit history graph only.