Module Irmin.Generic_key

"Generic key" stores are Irmin stores in which the backend may not be keyed directly by the hashes of stored values. See Key for more details.

module type S = sig ... end

module type KV = sig ... end

module type Maker = sig ... end

module type KV_maker = sig ... end

module type Maker_args = sig ... end

module Maker (X : Maker_args) : Maker with type ('h, 'v) contents_key = ('h, 'v) X.Contents_store.key and type 'h node_key = 'h X.Node_store.key and type 'h commit_key = 'h X.Commit_store.key