Module Irmin.Hash

Hashing functions.

Hash provides user-defined hash functions to digest serialized contents. Some backends might be parameterized by such hash functions, others might work with a fixed one (for instance, the Git format uses only SHA1).

A SHA1 implementation is available to pass to the backends.

Contents Hashing

module type S = sig ... end

Signature for hash values.

module type TYPED = sig ... end

Signature for typed hashes, where hash directly takes a value as argument and incremental hashing is not possible.

module Make = Hash.Make

Digestif hashes.

module Make_BLAKE2B = Hash.Make_BLAKE2B

module Make_BLAKE2S = Hash.Make_BLAKE2S

module SHA1 = Hash.SHA1

module RMD160 = Hash.RMD160

module SHA224 = Hash.SHA224

module SHA256 = Hash.SHA256

module SHA384 = Hash.SHA384

module SHA512 = Hash.SHA512

module BLAKE2B = Hash.BLAKE2B

module BLAKE2S = Hash.BLAKE2S

module V1 = Hash.V1

v1 serialisation

module Typed = Hash.Typed

Typed hashes.