Module Irmin.Info

Commit info are used to keep track of the origin of write operations in the stores. Info models the metadata associated with commit objects in Git.

module type S = sig ... end

module Default : S

type default = Default.t