Module Irmin.Metadata

Metadata defines metadata that is attached to contents but stored in nodes. For instance, the Git backend uses this to indicate the type of file (normal, executable or symlink).

module type S = sig ... end

module None : S with type t = unit

A metadata definition for systems that don't use metadata.