Module Irmin.Path

Store paths.

An Irmin store binds paths to user-defined contents. Paths are composed by basic elements, that we call steps. The following Path module provides functions to manipulate steps and paths.

Tree path handling.

module type S = sig ... end

Signature for path implementations.

module String_list : S with type step = string and type t = string list

An implementation of paths as string lists.