Module Private.Commit

Commit values represent the store history.

Every commit contains a list of predecessor commits, and the collection of commits form an acyclic directed graph.

Every commit also can contain an optional key, pointing to a node value. See the Node signature for more details on node values.

module type S = sig ... end

module Make : functor (K : Type.S) -> S with type hash = K.t

Make provides a simple implementation of commit values, parameterized by the commit and node keys K.

module V1 : functor (S : S) -> sig ... end

V1 serialisation.

module type STORE = sig ... end

STORE specifies the signature for commit stores.

module Store : functor (N : Node.STORE) -> functor (S : sig ... end) -> STORE with type 'a t = 'a N.t * 'a S.t and type key = S.key and type value = S.value and type Key.t = S.Key.t and module Val = S.Val

Store creates a new commit store.

module type HISTORY = sig ... end

History specifies the signature for commit history. The history is represented as a partial-order of commits and basic functions to search through that history are provided.

module History : functor (S : STORE) -> HISTORY with type 'a t = 'a S.t and type node = S.Node.key and type commit = S.key

Build a commit history.