Module Private.Commit

Commit values represent the store history.

Every commit contains a list of predecessor commits, and the collection of commits form an acyclic directed graph.

Every commit also can contain an optional key, pointing to a node value. See the Node signature for more details on node values.

module type S = sig ... end

module type Maker = sig ... end

module Make : Maker

Make provides a simple implementation of commit values, parameterized by the commit and node keys K.

module V1 (C : S) : sig ... end

V1 serialisation.

module type STORE = sig ... end

STORE specifies the signature for commit stores.

module Store (N : sig ... end) (C : sig ... end) : STORE with type 'a t = 'a N.t * 'a C.t and type key = C.key and type value = C.value and type Key.t = C.Key.t and module Val = C.Val

Store creates a new commit store.

module type HISTORY = sig ... end

History specifies the signature for commit history. The history is represented as a partial-order of commits and basic functions to search through that history are provided.

module History (C : STORE) : HISTORY with type 'a t = 'a C.t and type v = C.Val.t and type node = C.Node.key and type commit = C.key

Build a commit history.