Module Private.Node

Node provides functions to describe the graph-like structured values.

The node blocks form a labeled directed acyclic graph, labeled by steps: a list of steps defines a unique path from one node to an other.

Each node can point to user-defined contents values.

module Proof : sig ... end

module type S = sig ... end

module type Maker = sig ... end

module Make : Maker

Make provides a simple node implementation, parameterized by the contents and notes keys K, paths P and metadata M.

module V1 (N : S with type step = string) : sig ... end

v1 serialisation

module type STORE = sig ... end

STORE specifies the signature for node stores.

module Store (C : Contents.STORE) (P : Path.S) (M : sig ... end) (N : sig ... end) : STORE with type 'a t = 'a C.t * 'a N.t and type key = N.key and type value = N.value and module Path = P and module Metadata = M and type Key.t = N.key and module Val = N.Val

Store creates node stores.

module type GRAPH = sig ... end

Graph specifies the signature for node graphs. A node graph is a deterministic DAG, labeled by steps.

module Graph (N : STORE) : GRAPH with type 'a t = 'a N.t and type contents = N.Contents.key and type metadata = N.Metadata.t and type node = N.key and type step = N.Path.step and type path = N.Path.t

module No_metadata : sig ... end