Module Storage.Read_only

Parameters

module M : Make

module Key : Type.S

module Value : Type.S

Signature

Read-only stores

Read-only stores are store where it is only possible to read existing values.

type -'a t

The type for stores. The 'a phantom type carries information about the store mutability.

type key = Key.t

The type for keys.

type value = Value.t

The type for raw values.

val mem : [> Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

val v : Irmin__Conf.t -> Perms.read t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.