Module Irmin_data

This module contains data structure implementations used in the implementation of Irmin. It is exposed only for internal use, and does not provide a stable API.

module Fixed_size_string_set : sig ... end

Mutable sets of strings with a common length.