Module Irmin_mem

In-memory store.

A simple in-memory store, using hash tables. Once one of the functors below is instantiated to a module M, it has a unique shared hash-table: multiple invocation of M.create will see and manipulate the same contents.

val config : unit -> Irmin.config

Configuration values.

module Append_only : Irmin.APPEND_ONLY_STORE_MAKER

An in-memory store for append-only values.

module Atomic_write : Irmin.ATOMIC_WRITE_STORE_MAKER

An in-memory store with atomic-write guarantees.

module Make : Irmin.S_MAKER

Constructor for in-memory Irmin store.

module KV : functor (C : Irmin.Contents.S) -> Irmin.KV with type contents = C.t and type metadata = unit and type Private.Sync.endpoint = unit

Constructor for in-memory KV stores. Subtype of Irmin.KV_MAKER.