Module Irmin_mem

In-memory store.

A simple in-memory store, using hash tables. Once one of the functors below is instantiated to a module M, it has a unique shared hash-table: multiple invocation of M.create will see and manipulate the same contents.

module Conf : sig ... end

val config : unit -> Irmin.config

Configuration values.

module Append_only : Irmin.Append_only.Maker

An in-memory store for append-only values.

module Content_addressable : Irmin.Content_addressable.Maker

An in-memory store for content-addressable values.

module Atomic_write : Irmin.Atomic_write.Maker

An in-memory store with atomic-write guarantees.

module KV : Irmin.KV_maker with type endpoint = unit and type metadata = unit and type info = Irmin.Info.default

Constructor for in-memory KV stores.

Constructor for in-memory Irmin store.

include Irmin.Maker with type endpoint = unit

type endpoint = unit

include Irmin.Key.Store_spec.S with type ('h, _) contents_key = 'h with type 'h node_key = 'h with type 'h commit_key = 'h

type ('h, _) contents_key = 'h

type 'h node_key = 'h

type 'h commit_key = 'h

module Make (Schema : Irmin.Schema.S) : sig ... end