irmin-pack

The irmin-pack package provides a space-optimised, on-disk backend inspired by Git Packfiles.

irmin-pack is designed and developed for a specific use-case and requires more in-depth backend configuration from applications. It offers additional functionality through extensions to the core Irmin API, but users interested in a more general Irmin backend with simpler configuration are recommended to use other backends, such as Irmin_fs or Irmin_git.

See examples/irmin-pack for examples of how to configure and use some of the additional functionality.

The entry point of this library is the module Irmin_pack.

Implementations

irmin-pack.mem

The module Irmin_pack_mem is this library's entry point.

irmin-pack.unix

The module Irmin_pack_unix is this library's entry point.